+++ tile.cpp
@@ -37,6 +37,7 @@
 #include "options.hpp"
 #include "main.hpp"
 #include "write_json.hpp"
+#include "milo/dtoa_milo.h"
 
 extern "C" {
 #include "jsonpull/jsonpull.h"
@@ -72,7 +73,7 @@ bool draws_something(drawvec &geom) {
 	return false;
 }
 
-int metacmp(int m1, const std::vector<long long> &keys1, const std::vector<long long> &values1, char *stringpool1, int m2, const std::vector<long long> &keys2, const std::vector<long long> &values2, char *stringpool2);
+static int metacmp(int m1, const std::vector<long long> &keys1, const std::vector<long long> &values1, char *stringpool1, int m2, const std::vector<long long> &keys2, const std::vector<long long> &values2, char *stringpool2);
 int coalindexcmp(const struct coalesce *c1, const struct coalesce *c2);
 
 struct coalesce {
@@ -129,7 +130,38 @@ int coalcmp(const void *v1, const void *v2) {
 		}
 	}
 
-	return metacmp(c1->m, c1->keys, c1->values, c1->stringpool, c2->m, c2->keys, c2->values, c2->stringpool);
+	cmp = metacmp(c1->m, c1->keys, c1->values, c1->stringpool, c2->m, c2->keys, c2->values, c2->stringpool);
+	if (cmp != 0) {
+		return cmp;
+	}
+
+	if (c1->full_keys.size() < c2->full_keys.size()) {
+		return -1;
+	} else if (c1->full_keys.size() > c2->full_keys.size()) {
+		return 1;
+	}
+
+	for (size_t i = 0; i < c1->full_keys.size(); i++) {
+		if (c1->full_keys[i] < c1->full_keys[i]) {
+			return -1;
+		} else if (c1->full_keys[i] > c1->full_keys[i]) {
+			return 1;
+		}
+
+		if (c1->full_values[i].type < c1->full_values[i].type) {
+			return -1;
+		} else if (c1->full_values[i].type > c1->full_values[i].type) {
+			return 1;
+		}
+
+		if (c1->full_values[i].s < c1->full_values[i].s) {
+			return -1;
+		} else if (c1->full_values[i].s > c1->full_values[i].s) {
+			return 1;
+		}
+	}
+
+	return 0;
 }
 
 int coalindexcmp(const struct coalesce *c1, const struct coalesce *c2) {
@@ -174,7 +206,7 @@ void decode_meta(int m, std::vector<long long> const &metakeys, std::vector<long
 	}
 }
 
-int metacmp(int m1, const std::vector<long long> &keys1, const std::vector<long long> &values1, char *stringpool1, int m2, const std::vector<long long> &keys2, const std::vector<long long> &values2, char *stringpool2) {
+static int metacmp(int m1, const std::vector<long long> &keys1, const std::vector<long long> &values1, char *stringpool1, int m2, const std::vector<long long> &keys2, const std::vector<long long> &values2, char *stringpool2) {
 	int i;
 	for (i = 0; i < m1 && i < m2; i++) {
 		mvt_value key1 = retrieve_string(keys1[i], stringpool1, NULL);
@@ -342,6 +374,9 @@ struct partial {
 	unsigned long long id = 0;
 	bool has_id = 0;
 	ssize_t renamed = 0;
+	long long extent = 0;
+	long long clustered = 0;
+	std::set<std::string> need_tilestats;
 };
 
 struct partial_arg {
@@ -396,7 +431,14 @@ void *partial_feature_worker(void *v) {
 	std::vector<struct partial> *partials = a->partials;
 
 	for (size_t i = a->task; i < (*partials).size(); i += a->tasks) {
-		drawvec geom = (*partials)[i].geoms[0];  // XXX assumption of a single geometry at the beginning
+		drawvec geom;
+
+		for (size_t j = 0; j < (*partials)[i].geoms.size(); j++) {
+			for (size_t k = 0; k < (*partials)[i].geoms[j].size(); k++) {
+				geom.push_back((*partials)[i].geoms[j][k]);
+			}
+		}
+
 		(*partials)[i].geoms.clear();		 // avoid keeping two copies in memory
 		signed char t = (*partials)[i].t;
 		int z = (*partials)[i].z;
@@ -1165,6 +1207,7 @@ struct write_tile_args {
 	double fraction_out = 0;
 	const char *prefilter = NULL;
 	const char *postfilter = NULL;
+	std::map<std::string, attribute_op> const *attribute_accum = NULL;
 	bool still_dropping = false;
 	int wrote_zoom = 0;
 	size_t tiling_seg = 0;
@@ -1287,8 +1330,9 @@ serial_feature next_feature(FILE *geoms, long long *geompos_in, char *metabase, 
 		if (sf.tippecanoe_maxzoom != -1 && z > sf.tippecanoe_maxzoom) {
 			continue;
 		}
+
 		if (sf.tippecanoe_minzoom == -1 && z < sf.feature_minzoom) {
-			continue;
+			sf.dropped = true;
 		}
 
 		return sf;
@@ -1352,6 +1396,7 @@ void *run_prefilter(void *v) {
 		tmp_feature.geometry = to_feature(sf.geometry);
 		tmp_feature.id = sf.id;
 		tmp_feature.has_id = sf.has_id;
+		tmp_feature.dropped = sf.dropped;
 
 		// Offset from tile coordinates back to world coordinates
 		unsigned sx = 0, sy = 0;
@@ -1367,7 +1412,7 @@ void *run_prefilter(void *v) {
 		decode_meta(sf.m, sf.keys, sf.values, rpa->stringpool + rpa->pool_off[sf.segment], tmp_layer, tmp_feature);
 		tmp_layer.features.push_back(tmp_feature);
 
-		layer_to_geojson(rpa->prefilter_fp, tmp_layer, 0, 0, 0, false, true, false, sf.index, sf.seq, sf.extent, true);
+		layer_to_geojson(rpa->prefilter_fp, tmp_layer, 0, 0, 0, false, true, false, true, sf.index, sf.seq, sf.extent, true);
 	}
 
 	if (fclose(rpa->prefilter_fp) != 0) {
@@ -1411,6 +1456,144 @@ void add_tilestats(std::string const &layername, int z, std::vector<std::map<std
 	attrib.string = val.s;
 
 	add_to_file_keys(fk->second.file_keys, key, attrib);
+}
+
+struct accum_state {
+	double sum = 0;
+	double count = 0;
+};
+
+void preserve_attribute(attribute_op op, std::map<std::string, accum_state> &attribute_accum_state, serial_feature &, char *stringpool, long long *pool_off, std::string &key, serial_val &val, partial &p) {
+	if (p.need_tilestats.count(key) == 0) {
+		p.need_tilestats.insert(key);
+	}
+
+	// If the feature being merged into has this key as a metadata reference,
+	// promote it to a full_key so it can be modified
+
+	for (int i = 0; i < p.m; i++) {
+		if (strcmp(key.c_str(), stringpool + pool_off[p.segment] + p.keys[i] + 1) == 0) {
+			serial_val sv;
+			sv.s = stringpool + pool_off[p.segment] + p.values[i] + 1;
+			sv.type = (stringpool + pool_off[p.segment])[p.values[i]];
+
+			p.full_keys.push_back(key);
+			p.full_values.push_back(sv);
+
+			p.keys.erase(p.keys.begin() + i);
+			p.values.erase(p.values.begin() + i);
+			p.m--;
+
+			break;
+		}
+	}
+
+	for (size_t i = 0; i < p.full_keys.size(); i++) {
+		if (key == p.full_keys[i]) {
+			switch (op) {
+			case op_sum:
+				p.full_values[i].s = milo::dtoa_milo(atof(p.full_values[i].s.c_str()) + atof(val.s.c_str()));
+				p.full_values[i].type = mvt_double;
+				break;
+
+			case op_product:
+				p.full_values[i].s = milo::dtoa_milo(atof(p.full_values[i].s.c_str()) * atof(val.s.c_str()));
+				p.full_values[i].type = mvt_double;
+				break;
+
+			case op_max:
+				{
+					double existing = atof(p.full_values[i].s.c_str());
+					double maybe = atof(val.s.c_str());
+					if (maybe > existing) {
+						p.full_values[i].s = val.s.c_str();
+						p.full_values[i].type = mvt_double;
+					}
+				}
+				break;
+
+			case op_min:
+				{
+					double existing = atof(p.full_values[i].s.c_str());
+					double maybe = atof(val.s.c_str());
+					if (maybe < existing) {
+						p.full_values[i].s = val.s.c_str();
+						p.full_values[i].type = mvt_double;
+					}
+				}
+				break;
+
+			case op_mean:
+				{
+					auto state = attribute_accum_state.find(key);
+					if (state == attribute_accum_state.end()) {
+						accum_state s;
+						s.sum = atof(p.full_values[i].s.c_str()) + atof(val.s.c_str());
+						s.count = 2;
+						attribute_accum_state.insert(std::pair<std::string, accum_state>(key, s));
+
+						p.full_values[i].s = milo::dtoa_milo(s.sum / s.count);
+					} else {
+						state->second.sum += atof(val.s.c_str());
+						state->second.count += 1;
+
+						p.full_values[i].s = milo::dtoa_milo(state->second.sum / state->second.count);
+					}
+				}
+				break;
+
+			case op_concat:
+				p.full_values[i].s += val.s;
+				p.full_values[i].type = mvt_string;
+				break;
+
+			case op_comma:
+				p.full_values[i].s += std::string(",") + val.s;
+				p.full_values[i].type = mvt_string;
+				break;
+			}
+		}
+	}
+}
+
+void preserve_attributes(std::map<std::string, attribute_op> const *attribute_accum, std::map<std::string, accum_state> &attribute_accum_state, serial_feature &sf, char *stringpool, long long *pool_off, partial &p) {
+	for (size_t i = 0; i < sf.m; i++) {
+		std::string key = stringpool + pool_off[sf.segment] + sf.keys[i] + 1;
+
+		serial_val sv;
+		sv.type = (stringpool + pool_off[sf.segment])[sf.values[i]];
+		sv.s = stringpool + pool_off[sf.segment] + sf.values[i] + 1;
+
+		auto f = attribute_accum->find(key);
+		if (f != attribute_accum->end()) {
+			preserve_attribute(f->second, attribute_accum_state, sf, stringpool, pool_off, key, sv, p);
+		}
+	}
+	for (size_t i = 0; i < sf.full_keys.size(); i++) {
+		std::string key = sf.full_keys[i];
+		serial_val sv = sf.full_values[i];
+
+		auto f = attribute_accum->find(key);
+		if (f != attribute_accum->end()) {
+			preserve_attribute(f->second, attribute_accum_state, sf, stringpool, pool_off, key, sv, p);
+		}
+	}
+}
+
+bool find_partial(std::vector<partial> &partials, serial_feature &sf, ssize_t &out, std::vector<std::vector<std::string>> *layer_unmaps) {
+	for (size_t i = partials.size(); i > 0; i--) {
+		if (partials[i - 1].t == sf.t) {
+			std::string &layername1 = (*layer_unmaps)[partials[i - 1].segment][partials[i - 1].layer];
+			std::string &layername2 = (*layer_unmaps)[sf.segment][sf.layer];
+
+			if (layername1 == layername2) {
+				out = i - 1;
+				return true;
+			}
+		}
+	}
+
+	return false;
 }
 
 long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *stringpool, int z, unsigned tx, unsigned ty, int detail, int min_detail, sqlite3 *outdb, const char *outdir, int buffer, const char *fname, FILE **geomfile, int minzoom, int maxzoom, double todo, volatile long long *along, long long alongminus, double gamma, int child_shards, long long *meta_off, long long *pool_off, unsigned *initial_x, unsigned *initial_y, volatile int *running, double simplification, std::vector<std::map<std::string, layermap_entry>> *layermaps, std::vector<std::vector<std::string>> *layer_unmaps, size_t tiling_seg, size_t pass, size_t passes, unsigned long long mingap, long long minextent, double fraction, const char *prefilter, const char *postfilter, write_tile_args *arg) {
@@ -1457,7 +1640,6 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 		double scale = (double) (1LL << (64 - 2 * (z + 8)));
 		double gap = 0, density_gap = 0;
 		double spacing = 0;
-		size_t clustered = 0;
 
 		long long original_features = 0;
 		long long unclipped_features = 0;
@@ -1466,7 +1648,8 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 		std::map<std::string, std::vector<coalesce>> layers;
 		std::vector<unsigned long long> indices;
 		std::vector<long long> extents;
-		std::vector<serial_feature> coalesced_geometry;
+		std::map<std::string, accum_state> attribute_accum_state;
+		double coalesced_area = 0;
 
 		int within[child_shards];
 		long long geompos[child_shards];
@@ -1545,6 +1728,7 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 
 		while (1) {
 			serial_feature sf;
+			ssize_t which_partial = -1;
 
 			if (prefilter == NULL) {
 				sf = next_feature(geoms, geompos_in, metabase, meta_off, z, tx, ty, initial_x, initial_y, &original_features, &unclipped_features, nextzoom, maxzoom, minzoom, max_zoom_increment, pass, passes, along, alongminus, buffer, within, &first_time, geomfile, geompos, &oprogress, todo, fname, child_shards);
@@ -1556,66 +1740,53 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 				break;
 			}
 
-			if (gamma > 0) {
-				if (manage_gap(sf.index, &previndex, scale, gamma, &gap)) {
+			if (sf.dropped) {
+				if (find_partial(partials, sf, which_partial, layer_unmaps)) {
+					preserve_attributes(arg->attribute_accum, attribute_accum_state, sf, stringpool, pool_off, partials[which_partial]);
 					continue;
 				}
 			}
 
-			double coalesced_area = 0;
-			for (size_t i = 0; i < coalesced_geometry.size(); i++) {
-				if (coalesced_geometry[i].t == sf.t) {
-					coalesced_area += coalesced_geometry[i].extent;
+			if (gamma > 0) {
+				if (manage_gap(sf.index, &previndex, scale, gamma, &gap) && find_partial(partials, sf, which_partial, layer_unmaps)) {
+					preserve_attributes(arg->attribute_accum, attribute_accum_state, sf, stringpool, pool_off, partials[which_partial]);
+					continue;
 				}
 			}
 
 			if (additional[A_CLUSTER_DENSEST_AS_NEEDED] || cluster_distance != 0) {
 				indices.push_back(sf.index);
-				if (sf.index - merge_previndex < mingap) {
-					clustered++;
+				if ((sf.index < merge_previndex || sf.index - merge_previndex < mingap) && find_partial(partials, sf, which_partial, layer_unmaps)) {
+					partials[which_partial].clustered++;
+					preserve_attributes(arg->attribute_accum, attribute_accum_state, sf, stringpool, pool_off, partials[which_partial]);
 					continue;
-				} else {
-					if (clustered > 0) {
-						std::string layername = (*layer_unmaps)[sf.segment][sf.layer];
-						serial_val sv, sv2;
-
-						sf.full_keys.push_back("clustered");
-						sv.type = mvt_bool;
-						sv.s = "true";
-						sf.full_values.push_back(sv);
-
-						add_tilestats(layername, z, layermaps, tiling_seg, layer_unmaps, "clustered", sv);
-
-						sf.full_keys.push_back("point_count");
-						sv2.type = mvt_double;
-						sv2.s = std::to_string(clustered + 1);
-						sf.full_values.push_back(sv2);
-
-						add_tilestats(layername, z, layermaps, tiling_seg, layer_unmaps, "point_count", sv2);
-					}
-
-					clustered = 0;
 				}
 			} else if (additional[A_DROP_DENSEST_AS_NEEDED]) {
 				indices.push_back(sf.index);
-				if (sf.index - merge_previndex < mingap) {
+				if (sf.index - merge_previndex < mingap && find_partial(partials, sf, which_partial, layer_unmaps)) {
+					preserve_attributes(arg->attribute_accum, attribute_accum_state, sf, stringpool, pool_off, partials[which_partial]);
 					continue;
 				}
 			} else if (additional[A_COALESCE_DENSEST_AS_NEEDED]) {
 				indices.push_back(sf.index);
-				if (sf.index - merge_previndex < mingap) {
-					coalesced_geometry.push_back(sf);
+				if (sf.index - merge_previndex < mingap && find_partial(partials, sf, which_partial, layer_unmaps)) {
+					partials[which_partial].geoms.push_back(sf.geometry);
+					coalesced_area += sf.extent;
+					preserve_attributes(arg->attribute_accum, attribute_accum_state, sf, stringpool, pool_off, partials[which_partial]);
 					continue;
 				}
 			} else if (additional[A_DROP_SMALLEST_AS_NEEDED]) {
 				extents.push_back(sf.extent);
-				if (sf.extent + coalesced_area <= minextent && sf.t != VT_POINT) {
+				if (sf.extent + coalesced_area <= minextent && sf.t != VT_POINT && find_partial(partials, sf, which_partial, layer_unmaps)) {
+					preserve_attributes(arg->attribute_accum, attribute_accum_state, sf, stringpool, pool_off, partials[which_partial]);
 					continue;
 				}
 			} else if (additional[A_COALESCE_SMALLEST_AS_NEEDED]) {
 				extents.push_back(sf.extent);
-				if (sf.extent + coalesced_area <= minextent) {
-					coalesced_geometry.push_back(sf);
+				if (sf.extent + coalesced_area <= minextent && find_partial(partials, sf, which_partial, layer_unmaps)) {
+					partials[which_partial].geoms.push_back(sf.geometry);
+					coalesced_area += sf.extent;
+					preserve_attributes(arg->attribute_accum, attribute_accum_state, sf, stringpool, pool_off, partials[which_partial]);
 					continue;
 				}
 			}
@@ -1633,24 +1804,15 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 			}
 
 			fraction_accum += fraction;
-			if (fraction_accum < 1) {
+			if (fraction_accum < 1 && find_partial(partials, sf, which_partial, layer_unmaps)) {
 				if (additional[A_COALESCE_FRACTION_AS_NEEDED]) {
-					coalesced_geometry.push_back(sf);
+					partials[which_partial].geoms.push_back(sf.geometry);
+					coalesced_area += sf.extent;
 				}
+				preserve_attributes(arg->attribute_accum, attribute_accum_state, sf, stringpool, pool_off, partials[which_partial]);
 				continue;
 			}
 			fraction_accum -= 1;
-
-			if (coalesced_geometry.size() != 0) {
-				for (ssize_t i = coalesced_geometry.size() - 1; i >= 0; i--) {
-					if (coalesced_geometry[i].t == sf.t && coalesced_geometry[i].layer == sf.layer) {
-						for (size_t j = 0; j < coalesced_geometry[i].geometry.size(); j++) {
-							sf.geometry.push_back(coalesced_geometry[i].geometry[j]);
-						}
-						coalesced_geometry.erase(coalesced_geometry.begin() + i);
-					}
-				}
-			}
 
 			bool reduced = false;
 			if (sf.t == VT_POLYGON) {
@@ -1682,60 +1844,51 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 				p.has_id = sf.has_id;
 				p.index = sf.index;
 				p.renamed = -1;
+				p.extent = sf.extent;
+				p.clustered = 0;
 				partials.push_back(p);
 			}
 
 			merge_previndex = sf.index;
+			coalesced_area = 0;
 		}
 
-		// Attach any pieces that were waiting to be coalesced onto some features that did make it.
-		for (ssize_t i = (ssize_t) coalesced_geometry.size() - 1; i >= 0; i--) {
-			for (ssize_t j = partials.size() - 1; j >= 0; j--) {
-				if (partials[j].layer == coalesced_geometry[i].layer && partials[j].t == coalesced_geometry[i].t) {
-					for (size_t k = 0; k < coalesced_geometry[i].geometry.size(); k++) {
-						partials[j].geoms[0].push_back(coalesced_geometry[i].geometry[k]);
-					}
+		for (size_t i = 0; i < partials.size(); i++) {
+			partial &p = partials[i];
 
-					coalesced_geometry.erase(coalesced_geometry.begin() + i);
-					break;
-				}
+			if (p.clustered > 0) {
+				std::string layername = (*layer_unmaps)[p.segment][p.layer];
+				serial_val sv, sv2, sv3;
+
+				p.full_keys.push_back("clustered");
+				sv.type = mvt_bool;
+				sv.s = "true";
+				p.full_values.push_back(sv);
+
+				add_tilestats(layername, z, layermaps, tiling_seg, layer_unmaps, "clustered", sv);
+
+				p.full_keys.push_back("point_count");
+				sv2.type = mvt_double;
+				sv2.s = std::to_string(p.clustered + 1);
+				p.full_values.push_back(sv2);
+
+				add_tilestats(layername, z, layermaps, tiling_seg, layer_unmaps, "point_count", sv2);
+
+				p.full_keys.push_back("sqrt_point_count");
+				sv3.type = mvt_double;
+				sv3.s = std::to_string(round(100 * sqrt(p.clustered + 1)) / 100.0);
+				p.full_values.push_back(sv3);
+
+				add_tilestats(layername, z, layermaps, tiling_seg, layer_unmaps, "sqrt_point_count", sv3);
 			}
-		}
 
-		// Attach the leftover cluster count to the last feature that did make it
-		if (clustered > 0) {
-			if (partials.size() > 0) {
-				size_t n = partials.size() - 1;
+			if (p.need_tilestats.size() > 0) {
+				std::string layername = (*layer_unmaps)[p.segment][p.layer];
 
-				size_t i;
-				for (i = 0; i < partials[n].full_keys.size(); i++) {
-					if (partials[n].full_keys[i] == std::string("point_count")) {
-						break;
+				for (size_t j = 0; j < p.full_keys.size(); j++) {
+					if (p.need_tilestats.count(p.full_keys[j]) > 0) {
+						add_tilestats(layername, z, layermaps, tiling_seg, layer_unmaps, p.full_keys[j], p.full_values[j]);
 					}
-				}
-
-				std::string layername = (*layer_unmaps)[partials[n].segment][partials[n].layer];
-				if (i < partials[n].full_keys.size()) {
-					size_t sum = strtoul(partials[n].full_values[i].s.c_str(), NULL, 10) + clustered;
-					partials[n].full_values[i].s = std::to_string(sum);
-
-					add_tilestats(layername, z, layermaps, tiling_seg, layer_unmaps, "point_count", partials[n].full_values[i]);
-				} else {
-					serial_val sv, sv2;
-
-					partials[n].full_keys.push_back("clustered");
-					sv.type = mvt_bool;
-					sv.s = "true";
-					partials[n].full_values.push_back(sv);
-
-					add_tilestats(layername, z, layermaps, tiling_seg, layer_unmaps, "clustered", sv);
-
-					partials[n].full_keys.push_back("point_count");
-					sv2.type = mvt_double;
-					sv2.s = std::to_string(clustered + 1);
-					partials[n].full_values.push_back(sv2);
-
-					add_tilestats(layername, z, layermaps, tiling_seg, layer_unmaps, "point_count", sv2);
 				}
 			}
 		}
@@ -1962,10 +2115,17 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 							glow = 255;
 						}
 					}
+
 					mvt_value v;
 					v.type = mvt_sint;
 					v.numeric_value.sint_value = glow;
 					layer.tag(feature, "tippecanoe_feature_density", v);
+
+					serial_val sv;
+					sv.type = mvt_double;
+					sv.s = std::to_string(glow);
+
+					add_tilestats(layer.name, z, layermaps, tiling_seg, layer_unmaps, "tippecanoe_feature_density", sv);
 				}
 
 				layer.features.push_back(feature);
@@ -2027,11 +2187,15 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 					}
 					line_detail++;  // to keep it the same when the loop decrements it
 					continue;
-				} else if (additional[A_DROP_DENSEST_AS_NEEDED] || additional[A_COALESCE_DENSEST_AS_NEEDED] || additional[A_CLUSTER_DENSEST_AS_NEEDED]) {
+				} else if (mingap < ULONG_MAX && (additional[A_DROP_DENSEST_AS_NEEDED] || additional[A_COALESCE_DENSEST_AS_NEEDED] || additional[A_CLUSTER_DENSEST_AS_NEEDED])) {
 					mingap_fraction = mingap_fraction * max_tile_features / totalsize * 0.90;
 					unsigned long long mg = choose_mingap(indices, mingap_fraction);
 					if (mg <= mingap) {
 						mg = (mingap + 1) * 1.5;
+
+						if (mg <= mingap) {
+							mg = ULONG_MAX;
+						}
 					}
 					mingap = mg;
 					if (mingap > arg->mingap_out) {
@@ -2058,7 +2222,10 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 						line_detail++;
 						continue;
 					}
-				} else if (prevent[P_DYNAMIC_DROP] || additional[A_DROP_FRACTION_AS_NEEDED] || additional[A_COALESCE_FRACTION_AS_NEEDED]) {
+				} else if (totalsize > layers.size() && (prevent[P_DYNAMIC_DROP] || additional[A_DROP_FRACTION_AS_NEEDED] || additional[A_COALESCE_FRACTION_AS_NEEDED])) {
+					// The 95% is a guess to avoid too many retries
+					// and probably actually varies based on how much duplicated metadata there is
+
 					fraction = fraction * max_tile_features / totalsize * 0.95;
 					if (!quiet) {
 						fprintf(stderr, "Going to try keeping %0.2f%% of the features to make it fit\n", fraction * 100);
@@ -2111,11 +2278,15 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 						fprintf(stderr, "Going to try gamma of %0.3f to make it fit\n", gamma);
 					}
 					line_detail++;  // to keep it the same when the loop decrements it
-				} else if (additional[A_DROP_DENSEST_AS_NEEDED] || additional[A_COALESCE_DENSEST_AS_NEEDED] || additional[A_CLUSTER_DENSEST_AS_NEEDED]) {
+				} else if (mingap < ULONG_MAX && (additional[A_DROP_DENSEST_AS_NEEDED] || additional[A_COALESCE_DENSEST_AS_NEEDED] || additional[A_CLUSTER_DENSEST_AS_NEEDED])) {
 					mingap_fraction = mingap_fraction * max_tile_size / compressed.size() * 0.90;
 					unsigned long long mg = choose_mingap(indices, mingap_fraction);
 					if (mg <= mingap) {
-						mg = mingap * 1.5;
+						mg = (mingap + 1) * 1.5;
+
+						if (mg <= mingap) {
+							mg = ULONG_MAX;
+						}
 					}
 					mingap = mg;
 					if (mingap > arg->mingap_out) {
@@ -2141,7 +2312,7 @@ long long write_tile(FILE *geoms, long long *geompos_in, char *metabase, char *s
 						line_detail++;
 						continue;
 					}
-				} else if (prevent[P_DYNAMIC_DROP] || additional[A_DROP_FRACTION_AS_NEEDED] || additional[A_COALESCE_FRACTION_AS_NEEDED]) {
+				} else if (totalsize > layers.size() && (prevent[P_DYNAMIC_DROP] || additional[A_DROP_FRACTION_AS_NEEDED] || additional[A_COALESCE_FRACTION_AS_NEEDED])) {
 					// The 95% is a guess to avoid too many retries
 					// and probably actually varies based on how much duplicated metadata there is
 
@@ -2295,7 +2466,7 @@ void *run_thread(void *vargs) {
 	return NULL;
 }
 
-int traverse_zooms(int *geomfd, off_t *geom_size, char *metabase, char *stringpool, unsigned *midx, unsigned *midy, int &maxzoom, int minzoom, sqlite3 *outdb, const char *outdir, int buffer, const char *fname, const char *tmpdir, double gamma, int full_detail, int low_detail, int min_detail, long long *meta_off, long long *pool_off, unsigned *initial_x, unsigned *initial_y, double simplification, std::vector<std::map<std::string, layermap_entry>> &layermaps, const char *prefilter, const char *postfilter) {
+int traverse_zooms(int *geomfd, off_t *geom_size, char *metabase, char *stringpool, unsigned *midx, unsigned *midy, int &maxzoom, int minzoom, sqlite3 *outdb, const char *outdir, int buffer, const char *fname, const char *tmpdir, double gamma, int full_detail, int low_detail, int min_detail, long long *meta_off, long long *pool_off, unsigned *initial_x, unsigned *initial_y, double simplification, std::vector<std::map<std::string, layermap_entry>> &layermaps, const char *prefilter, const char *postfilter, std::map<std::string, attribute_op> const *attribute_accum) {
 	// The existing layermaps are one table per input thread.
 	// We need to add another one per *tiling* thread so that it can be
 	// safely changed during tiling.
@@ -2481,6 +2652,7 @@ int traverse_zooms(int *geomfd, off_t *geom_size, char *metabase, char *stringpo
 				args[thread].tiling_seg = thread + layermaps_off;
 				args[thread].prefilter = prefilter;
 				args[thread].postfilter = postfilter;
+				args[thread].attribute_accum = attribute_accum;
 
 				args[thread].tasks = dispatches[thread].tasks;
 				args[thread].running = &running;